Extract `TableClickBehaviorView` component
