+++ frontend/src/metabase/dashboard/components/ClickBehaviorSidebar/ClickBehaviorSidebar.jsx
@@ -1,14 +1,12 @@
 /* eslint-disable react/prop-types */
 import React from "react";
-import { t, jt } from "ttag";
+import { jt } from "ttag";
 import { getIn } from "icepick";
-import _ from "underscore";
 
 import Icon from "metabase/components/Icon";
 
 import { color } from "metabase/lib/colors";
 import {
-  hasActionsMenu,
   isTableDisplay,
   clickBehaviorIsValid,
 } from "metabase/lib/click-behavior";
@@ -18,9 +16,9 @@ import Sidebar from "metabase/dashboard/components/Sidebar";
 
 import { clickBehaviorOptions, getClickBehaviorOptionName } from "./utils";
 import ActionOptions from "./ActionOptions";
-import Column from "./Column";
 import CrossfilterOptions from "./CrossfilterOptions";
 import LinkOptions from "./LinkOptions";
+import TableClickBehaviorView from "./TableClickBehaviorView";
 import TypeSelector from "./TypeSelector";
 import { SidebarItemWrapper } from "./SidebarItem";
 import {
@@ -155,59 +153,17 @@ class ClickBehaviorSidebar extends React.Component {
 
     if (isTableDisplay(dashcard) && selectedColumn == null) {
       return (
-        <Sidebar
-          onClose={hideClickBehaviorSidebar}
+        <TableClickBehaviorView
+          columns={this.getColumns()}
+          dashcard={dashcard}
+          getClickBehaviorForColumn={column =>
+            this.getClickBehaviorForColumn(this.props, column)
+          }
+          canClose={clickBehaviorIsValid(clickBehavior)}
+          onColumnClick={this.setSelectedColumn}
           onCancel={this.handleCancel}
-          closeIsDisabled={!clickBehaviorIsValid(clickBehavior)}
-        >
-          <SidebarHeader>
-            <Heading className="text-paragraph">{t`On-click behavior for each column`}</Heading>
-          </SidebarHeader>
-          <div>
-            {_.chain(this.getColumns())
-              .map(column => ({
-                column,
-                clickBehavior: this.getClickBehaviorForColumn(
-                  this.props,
-                  column,
-                ),
-              }))
-              .groupBy(({ clickBehavior }) => {
-                const { type = "actionMenu" } = clickBehavior || {};
-                return type;
-              })
-              .pairs()
-              .sortBy(([linkType]) =>
-                ["link", "crossfilter", "actionMenu"].indexOf(linkType),
-              )
-              .map(([linkType, columnsWithClickBehavior]) => (
-                <div key={linkType} className="mb2 px4">
-                  <h5 className="text-uppercase text-medium my1">
-                    {
-                      {
-                        link: t`Go to custom destination`,
-                        crossfilter: t`Update a dashboard filter`,
-                        actionMenu: hasActionsMenu(dashcard)
-                          ? t`Open the actions menu`
-                          : t`Do nothing`,
-                      }[linkType]
-                    }
-                  </h5>
-                  {columnsWithClickBehavior.map(
-                    ({ column, clickBehavior }, index) => (
-                      <Column
-                        key={index}
-                        column={column}
-                        clickBehavior={clickBehavior}
-                        onClick={() => this.setSelectedColumn(column)}
-                      />
-                    ),
-                  )}
-                </div>
-              ))
-              .value()}
-          </div>
-        </Sidebar>
+          onClose={hideClickBehaviorSidebar}
+        />
       );
     }
